create slash command for ban
