+++ resources/commands-form/ban.js
@@ -1,0 +1,13 @@
+const { SlashCommandBuilder } = require('discord.js');
+
+const slashBan = new SlashCommandBuilder()
+    .setName('ban')
+    .setDescription('Ban a user from Squire')
+    .addStringOption(option => 
+        option.setName('userID')
+            .setDescription('The userID you want to ban')
+            .setRequired(true)
+    )
+
+
+module.exports = slashBan
